refactor: extract shared dashboard helpers to dedupe duplicates
- Move annotation numeric value SQL expression to shared expressions.py
  helper used by annotation_graph.py and dashboard.py
- Extract chart formatting helpers (escapeHtml, getSeriesAverage,
  getAutoDecimals, getSuggestedUnitConfig, formatValueWithConfig) to
  widgetUtils.js shared by WidgetChart.jsx and WidgetEditorView.jsx
