+++ frontend/src/sections/dashboards/WidgetEditorView.jsx
@@ -57,15 +57,14 @@ import { useSnackbar } from "src/components/snackbar";
 import { format } from "date-fns";
 import CustomDateRangePicker from "src/components/custom-datepicker/DatePicker";
 
-const escapeHtml = (str) => {
-  if (typeof str !== "string") return str;
-  return str
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#039;");
-};
+import {
+  DEFAULT_DECIMALS,
+  escapeHtml,
+  formatValueWithConfig,
+  getAutoDecimals,
+  getSeriesAverage,
+  getSuggestedUnitConfig,
+} from "./widgetUtils";
 
 const escapeCsvField = (field) => {
   const str = String(field ?? "");
@@ -73,86 +72,6 @@ const escapeCsvField = (field) => {
     return `"${str.replace(/"/g, '""')}"`;
   }
   return str;
-};
-
-const DEFAULT_DECIMALS = 2;
-
-const getSeriesAverage = (points = []) => {
-  let total = 0;
-  let count = 0;
-  for (const pt of points) {
-    if (pt?.y == null) continue;
-    const y = Number(pt.y);
-    if (!Number.isFinite(y)) continue;
-    total += y;
-    count += 1;
-  }
-  return count > 0 ? total / count : null;
-};
-
-const getAutoDecimals = (series = []) => {
-  let minAbs = Infinity;
-  for (const s of series) {
-    for (const pt of s.data || []) {
-      const raw = typeof pt === "number" ? pt : pt?.y;
-      const value = Number(raw);
-      if (!Number.isFinite(value)) continue;
-      const abs = Math.abs(value);
-      if (abs > 0 && abs < minAbs) minAbs = abs;
-    }
-  }
-  if (minAbs === Infinity || minAbs >= 0.01) return DEFAULT_DECIMALS;
-  if (minAbs >= 0.001) return 3;
-  return 4;
-};
-
-const UNIT_LESS_AGGREGATIONS = new Set([
-  "count",
-  "count_distinct",
-  "pass_count",
-  "fail_count",
-]);
-
-const getSuggestedUnitConfig = (metricConfigs = []) => {
-  if (
-    metricConfigs.some((metric) =>
-      UNIT_LESS_AGGREGATIONS.has(metric?.aggregation),
-    )
-  ) {
-    return { unit: "", prefixSuffix: "prefix" };
-  }
-  const uniqueUnits = [
-    ...new Set(metricConfigs.map((metric) => metric?.unit).filter(Boolean)),
-  ];
-  if (uniqueUnits.length !== 1) {
-    return { unit: "", prefixSuffix: "prefix" };
-  }
-  const [unit] = uniqueUnits;
-  if (unit === "$") return { unit, prefixSuffix: "prefix" };
-  if (unit === "%") return { unit, prefixSuffix: "suffix" };
-  return { unit: "", prefixSuffix: "prefix" };
-};
-
-const formatValueWithConfig = (
-  val,
-  cfg,
-  { fallbackDecimals = DEFAULT_DECIMALS, includeUnit = true } = {},
-) => {
-  if (val == null) return "-";
-  const num = Number(val);
-  if (!Number.isFinite(num)) return "-";
-  const dec = Math.max(0, Math.min(6, cfg?.decimals ?? fallbackDecimals));
-  const unit = includeUnit ? cfg?.unit || "" : "";
-  const prefixSuffix = cfg?.prefixSuffix || "prefix";
-  let str;
-  if (Boolean(cfg?.abbreviation ?? true) && Math.abs(num) >= 1000000) {
-    str = `${(num / 1000000).toFixed(dec)}M`;
-  } else if (Boolean(cfg?.abbreviation ?? true) && Math.abs(num) >= 1000) {
-    str = `${(num / 1000).toFixed(dec)}K`;
-  } else {
-    str = num.toFixed(dec);
-  }
-  return prefixSuffix === "suffix" ? `${str}${unit}` : `${unit}${str}`;
 };
 
 const TIME_PRESETS = [
